Upgrade to babel-7 (#59)
This upgrades babel to version 7. It also upgrades the various
plugins that we use to their version 7 versions. Of note, the
`rollup-plugin-babel` was upgraded to its newest version which
uses babel-7 as well. As part of that upgrade, I changed to rollup
build config to just use the `.babelrc` located in the project
root, rather than maintaining a duplicate copy of it in the rollup
config. Further, rollup now provides the babel `external-helpers`
functionality, so we do not need to include it ourselves in our
babel config, meaning it can be removed from this project.